+++ geochemistrypi/data_mining/process/detect.py
@@ -4,16 +4,16 @@ import os
 import pandas as pd
 
 from ..constants import MLFLOW_ARTIFACT_DATA_PATH
-from ..model.detection import AbnormalDetectionWorkflowBase, IsolationForestAbnormalDetection, LocalOutlierFactorAbnormalDetection
+from ..model.detection import AnomalyDetectionWorkflowBase, IsolationForestAnomalyDetection, LocalOutlierFactorAnomalyDetection
 from ._base import ModelSelectionBase
 
 
-class AbnormalDetectionModelSelection(ModelSelectionBase):
-    """Simulate the normal way of invoking scikit-learn abnormal detection algorithms."""
+class AnomalyDetectionModelSelection(ModelSelectionBase):
+    """Simulate the normal way of invoking scikit-learn anomaly detection algorithms."""
 
     def __init__(self, model_name: str) -> None:
         self.model_name = model_name
-        self.ad_workflow = AbnormalDetectionWorkflowBase()
+        self.ad_workflow = AnomalyDetectionWorkflowBase()
         self.transformer_config = {}
 
     def activate(
@@ -31,8 +31,8 @@ class AbnormalDetectionModelSelection(ModelSelectionBase):
 
         # Model option
         if self.model_name == "Isolation Forest":
-            hyper_parameters = IsolationForestAbnormalDetection.manual_hyper_parameters()
-            self.ad_workflow = IsolationForestAbnormalDetection(
+            hyper_parameters = IsolationForestAnomalyDetection.manual_hyper_parameters()
+            self.ad_workflow = IsolationForestAnomalyDetection(
                 n_estimators=hyper_parameters["n_estimators"],
                 contamination=hyper_parameters["contamination"],
                 max_features=hyper_parameters["max_features"],
@@ -41,8 +41,8 @@ class AbnormalDetectionModelSelection(ModelSelectionBase):
             )
 
         if self.model_name == "Local Outlier Factor":
-            hyper_parameters = LocalOutlierFactorAbnormalDetection.manual_hyper_parameters()
-            self.ad_workflow = LocalOutlierFactorAbnormalDetection(
+            hyper_parameters = LocalOutlierFactorAnomalyDetection.manual_hyper_parameters()
+            self.ad_workflow = LocalOutlierFactorAnomalyDetection(
                 n_neighbors=hyper_parameters["n_neighbors"],
                 contamination=hyper_parameters["contamination"],
                 leaf_size=hyper_parameters["leaf_size"],
@@ -55,23 +55,23 @@ class AbnormalDetectionModelSelection(ModelSelectionBase):
         # Use Scikit-learn style API to process input data
         self.ad_workflow.fit(X)
         y_predict = self.ad_workflow.predict(X)
-        X_abnormal_detection, X_normal, X_abnormal = self.ad_workflow._detect_data(X, y_predict)
+        X_anomaly_detection, X_normal, X_anomaly = self.ad_workflow._detect_data(X, y_predict)
 
         self.ad_workflow.data_upload(X=X, y=y, X_train=X_train, X_test=X_test, y_train=y_train, y_test=y_test)
 
         # Save the model hyper-parameters
         self.ad_workflow.save_hyper_parameters(hyper_parameters, self.model_name, os.getenv("GEOPI_OUTPUT_PARAMETERS_PATH"))
 
-        # Common components for every abnormal detection algorithm
+        # Common components for every anomaly detection algorithm
         self.ad_workflow.common_components()
 
         # special components of different algorithms
         self.ad_workflow.special_components()
 
-        # Save abnormal detection result
-        self.ad_workflow.data_save(X_abnormal_detection, "X Abnormal Detection", os.getenv("GEOPI_OUTPUT_ARTIFACTS_DATA_PATH"), MLFLOW_ARTIFACT_DATA_PATH, "Abnormal Detection Data")
+        # Save anomaly detection result
+        self.ad_workflow.data_save(X_anomaly_detection, "X Anomaly Detection", os.getenv("GEOPI_OUTPUT_ARTIFACTS_DATA_PATH"), MLFLOW_ARTIFACT_DATA_PATH, "Anomaly Detection Data")
         self.ad_workflow.data_save(X_normal, "X Normal", os.getenv("GEOPI_OUTPUT_ARTIFACTS_DATA_PATH"), MLFLOW_ARTIFACT_DATA_PATH, "Normal Data")
-        self.ad_workflow.data_save(X_abnormal, "X Abnormal", os.getenv("GEOPI_OUTPUT_ARTIFACTS_DATA_PATH"), MLFLOW_ARTIFACT_DATA_PATH, "Abnormal Data")
+        self.ad_workflow.data_save(X_anomaly, "X Anomaly", os.getenv("GEOPI_OUTPUT_ARTIFACTS_DATA_PATH"), MLFLOW_ARTIFACT_DATA_PATH, "Anomaly Data")
 
         # Save the trained model
         self.ad_workflow.model_save()